requirements.txt
Pillow==8.1.0
opencv-python==4.5.1.48
configparser==5.0.1
WSGIserver==1.3